#227 Reorganize the KSP layer
Split the 614-line KspUtils.kt grab-bag by concern (CreatorResolution, KspVisibility,
KspTypeRendering, KspEntityPredicates, KspNaming), centralize the scattered generated-name
suffix literals into LirpGenNames, collapse RegistryBase's duplicated Class.forName accessor
lookup into a generic KspAccessorLoader, and collapse the 9 near-identical *ProcessorProvider
files onto a shared LirpProcessorProvider base. Behavior-preserving.

KspAccessorLoader fails loud (error()) when a resolved generated accessor cannot be
instantiated, instead of silently caching an empty result and re-running the failing
reflection on every call.
