+++ lirp-ksp/src/main/kotlin/net/transgressoft/lirp/ksp/TableDefProcessorProvider.kt
@@ -17,15 +17,15 @@
 
 package net.transgressoft.lirp.ksp
 
+import com.google.devtools.ksp.processing.CodeGenerator
+import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.SymbolProcessor
-import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
-import com.google.devtools.ksp.processing.SymbolProcessorProvider
 
 /**
- * KSP entry point for the LIRP table descriptor processor, registered via `META-INF/services`.
+ * KSP entry point for [TableDefProcessor], registered via `META-INF/services`.
  */
-class TableDefProcessorProvider : SymbolProcessorProvider {
+class TableDefProcessorProvider : LirpProcessorProvider() {
 
-    override fun create(environment: SymbolProcessorEnvironment): SymbolProcessor =
-        TableDefProcessor(environment.codeGenerator, environment.logger)
+    override fun createProcessor(codeGenerator: CodeGenerator, logger: KSPLogger): SymbolProcessor =
+        TableDefProcessor(codeGenerator, logger)
 }
